- change 'omp for' iteration space to 'unsigned long long' - add basic support for final tasks, included tasks, merged tasks - add inline 'omp for' variants returning begin/end instead of first/nb - add support for data handles in omp_parallel_region/omp_task_region - update test cases

+++ src/util/openmp_runtime_support.h
@@ -278,12 +278,12 @@ LIST_TYPE(starpu_omp_thread,
 struct starpu_omp_loop
 {
 	int id;
-	unsigned long next_iteration;
+	unsigned long long next_iteration;
 	int nb_completed_threads;
 	struct starpu_omp_loop *next_loop;
 	struct _starpu_spinlock ordered_lock;
 	struct starpu_omp_condition ordered_cond;
-	unsigned long ordered_iteration;
+	unsigned long long ordered_iteration;
 };
 
 struct starpu_omp_region
